# docs/swagger.yaml
basePath: /
definitions:
  dangbamgong-backend_internal_domain.ErrorCode:
    enum:
    - BAD_REQUEST
    - NOT_FOUND
    - INTERNAL_SERVER_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - CONFLICT
    - SERVICE_UNAVAILABLE
    - INVALID_TOKEN
    - INVALID_NICKNAME
    - NICKNAME_ALREADY_SET
    - USER_NOT_FOUND
    - ALREADY_BLOCKED
    - NOT_BLOCKED
    - ALREADY_IN_VOID
    - NOT_IN_VOID
    - TOO_MANY_ACTIVITIES
    - INVALID_ACTIVITY_NAME
    - ACTIVITY_ALREADY_EXISTS
    - ACTIVITY_NOT_FOUND
    - NOTIFICATION_NOT_FOUND
    - ALREADY_FRIENDS
    - REQUEST_ALREADY_SENT
    - BLOCKED
    - REQUEST_NOT_FOUND
    - REQUEST_NOT_PENDING
    - NOT_FRIENDS
    - FRIEND_NOT_IN_VOID
    - INVALID_REQUEST_TYPE
    type: string
    x-enum-varnames:
    - ErrBadRequest
    - ErrNotFound
    - ErrInternalServer
    - ErrUnauthorized
    - ErrForbidden
    - ErrConflict
    - ErrServiceUnavailable
    - ErrInvalidToken
    - ErrInvalidNickname
    - ErrNicknameAlreadySet
    - ErrUserNotFound
    - ErrAlreadyBlocked
    - ErrNotBlocked
    - ErrAlreadyInVoid
    - ErrNotInVoid
    - ErrTooManyActivities
    - ErrInvalidActivityName
    - ErrActivityAlreadyExists
    - ErrActivityNotFound
    - ErrNotificationNotFound
    - ErrAlreadyFriends
    - ErrRequestAlreadySent
    - ErrBlocked
    - ErrRequestNotFound
    - ErrRequestNotPending
    - ErrNotFriends
    - ErrFriendNotInVoid
    - ErrInvalidRequestType
  dangbamgong-backend_internal_dto.ActivityItem:
    properties:
      id:
        type: string
      lastUsedAt:
        type: string
      name:
        type: string
      usageCount:
        type: integer
    type: object
  dangbamgong-backend_internal_dto.ActivityListResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.ActivityItem'
        type: array
    type: object
...
  dangbamgong-backend_internal_dto.CreateActivityRequest:
    properties:
      name:
        maxLength: 10
        minLength: 1
        type: string
    required:
    - name
    type: object
  dangbamgong-backend_internal_dto.CreateActivityResponse:
    properties:
      id:
        type: string
      lastUsedAt:
        type: string
      name:
        type: string
      usageCount:
        type: integer
    type: object
...
  dangbamgong-backend_internal_dto.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/dangbamgong-backend_internal_domain.ErrorCode'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.LoginRequest:
    properties:
      appleRefreshToken:
        type: string
      idToken:
        type: string
      provider:
        enum:
        - GOOGLE
        - KAKAO
        - APPLE
        type: string
    required:
    - idToken
    - provider
    type: object
  dangbamgong-backend_internal_dto.LoginResponse:
    properties:
      accessToken:
        type: string
      isNewUser:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.RegisterDeviceRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  dangbamgong-backend_internal_dto.Response-any:
    properties:
      data: {}
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ActivityListResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.ActivityListResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_CreateActivityResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.CreateActivityResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.LoginResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_SetNicknameResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.SetNicknameResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.SetNicknameRequest:
    properties:
      nickname:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - nickname
    type: object
  dangbamgong-backend_internal_dto.SetNicknameResponse:
    properties:
      nickname:
        type: string
    type: object
  dangbamgong-backend_internal_dto.TestLoginRequest:
    properties:
      socialId:
        type: string
    required:
    - socialId
    type: object
...
  dangbamgong-backend_internal_dto.UpdateActivityRequest:
    properties:
      name:
        maxLength: 10
        minLength: 1
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 밤의 공백 시간 측정 소셜 앱 API
  title: 당밤공 API
  version: "1.0"
paths:
...
  /activities:
    get:
      description: 유저의 활동 목록을 사용 빈도순으로 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ActivityListResponse'
      security:
      - BearerAuth: []
      summary: 활동 목록 조회
      tags:
      - Activities
    post:
      consumes:
      - application/json
      description: 새로운 활동을 생성합니다. 이름은 1-10자, 중복 불가.
      parameters:
      - description: 활동 이름
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_CreateActivityResponse'
        "400":
          description: INVALID_ACTIVITY_NAME
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "409":
          description: ACTIVITY_ALREADY_EXISTS
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 활동 생성
      tags:
      - Activities
  /activities/{activity_id}:
    delete:
      description: 활동을 삭제합니다
      parameters:
      - description: 활동 ID
        in: path
        name: activity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "404":
          description: ACTIVITY_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 활동 삭제
      tags:
      - Activities
    patch:
      consumes:
      - application/json
      description: 기존 활동의 이름을 수정합니다
      parameters:
      - description: 활동 ID
        in: path
        name: activity_id
        required: true
        type: string
      - description: 새 이름
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.UpdateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: INVALID_ACTIVITY_NAME
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "404":
          description: ACTIVITY_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "409":
          description: ACTIVITY_ALREADY_EXISTS
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 활동 이름 수정
      tags:
      - Activities
  /auth/login:
    post:
      consumes:
      - application/json
      description: Google/Kakao/Apple 소셜 로그인을 처리합니다. 신규 유저인 경우 자동 가입됩니다.
      parameters:
      - description: 소셜 로그인 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      summary: 소셜 로그인
      tags:
      - Auth
  /auth/login/test:
    post:
      consumes:
      - application/json
      description: 개발용 테스트 로그인. socialId로 직접 로그인합니다.
      parameters:
      - description: 테스트 로그인 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.TestLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_LoginResponse'
      summary: 테스트 로그인
      tags:
      - Auth
  /auth/nickname:
    post:
      consumes:
      - application/json
      description: 신규 가입 후 최초 닉네임을 설정합니다. 이미 설정된 경우 실패합니다.
      parameters:
      - description: 닉네임 (3-15자)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.SetNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_SetNicknameResponse'
        "400":
          description: INVALID_NICKNAME
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "409":
          description: NICKNAME_ALREADY_SET
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 닉네임 설정
      tags:
      - Auth
  /auth/withdraw:
    delete:
      description: 계정을 삭제합니다. (미완성)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
      security:
      - BearerAuth: []
      summary: 회원 탈퇴
      tags:
      - Auth
  /devices/token:
    delete:
      consumes:
      - application/json
      description: 등록된 디바이스 토큰을 삭제합니다
      parameters:
      - description: 삭제할 토큰
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 디바이스 토큰 삭제
      tags:
      - Devices
    put:
      consumes:
      - application/json
      description: APNs 푸시 알림용 디바이스 토큰을 등록(또는 갱신)합니다
      parameters:
      - description: 디바이스 토큰
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 디바이스 토큰 등록
      tags:
      - Devices
...
securityDefinitions:
  BearerAuth:
    description: Bearer {JWT 토큰}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"